+++ test/raylib/test_runner.clj
@@ -45,6 +45,8 @@
                      raylib.scenes.bullets-test
                      raylib.scenes.collision-test
                      raylib.scenes.dashed-test
+                     raylib.scenes.multitouch-test
+                     raylib.scenes.analog-test
                      raylib.easings-test
                      raylib.scenes.clock-test
                      raylib.scenes.easings-test